+++ lib/lightning/invocation.ex
@@ -356,9 +356,38 @@ defmodule Lightning.Invocation do
     end)
   end
 
+  def filter_run_body_and_logs_where(search_term, searchfors)
+      when search_term in ["", nil] or searchfors == [] do
+    dynamic(true)
+  end
+
+  def filter_run_body_and_logs_where(search_term, searchfors) do
+    Enum.reduce(searchfors, dynamic(false), fn
+      :log, query ->
+        dynamic(
+          [runs: r],
+          ^query or
+            fragment("cast(?  as VARCHAR) ilike ?", r.log, ^"%#{search_term}%")
+        )
+
+      :body, query ->
+        dynamic(
+          [input: i],
+          ^query or
+            fragment("cast(?  as VARCHAR) ilike ?", i.body, ^"%#{search_term}%")
+        )
+
+      _, query ->
+        # Not a where parameter
+        query
+    end)
+  end
+
   def list_work_orders_for_project_query(
         %Project{id: project_id},
         status: status,
+        searchfors: searchfors,
+        search_term: search_term,
         workflow_id: workflow_id,
         date_after: date_after,
         date_before: date_before
@@ -368,10 +397,14 @@ defmodule Lightning.Invocation do
       from(r in Lightning.Invocation.Run,
         as: :runs,
         join: j in assoc(r, :job),
+        join: d in assoc(r, :input_dataclip),
+        as: :input,
         order_by: [asc: r.finished_at],
+        where: r.input_dataclip_id == d.id,
         where: ^filter_run_status_where(status),
         where: ^filter_run_started_after_where(date_after),
         where: ^filter_run_started_before_where(date_before),
+        where: ^filter_run_body_and_logs_where(search_term, searchfors),
         preload: [
           job:
             ^from(job in Lightning.Jobs.Job,
@@ -400,12 +433,16 @@ defmodule Lightning.Invocation do
       join: att in assoc(wo, :attempts),
       join: r in assoc(att, :runs),
       as: :runs,
+      join: d in assoc(r, :input_dataclip),
+      as: :input,
       where: w.project_id == ^project_id,
+      where: re.dataclip_id == d.id,
       where: ^filter_workflow_where(workflow_id),
       where: ^filter_run_status_where(status),
       where: ^filter_run_started_after_where(date_after),
       where: ^filter_run_started_before_where(date_before),
-      distinct: true,
+      where: ^filter_run_body_and_logs_where(search_term, searchfors),
+      # distinct: true,
       order_by: [desc_nulls_first: r.finished_at],
       preload: [
         reason:
@@ -430,7 +467,8 @@ defmodule Lightning.Invocation do
   def list_work_orders_for_project(%Project{} = project, filter, params) do
     list_work_orders_for_project_query(project, filter)
     |> Repo.paginate(params)
-    |> find_uniq_wo()
+
+    # |> find_uniq_wo()
   end
 
   def find_uniq_wo(page) do
@@ -442,6 +480,8 @@ defmodule Lightning.Invocation do
       project,
       [
         status: [:success, :failure, :timeout, :crash, :pending],
+        searchfors: [],
+        search_term: "",
         workflow_id: "",
         date_after: "",
         date_before: ""